requirements.txt
bleach==1.5.0
certifi==2017.4.17
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
crontab==0.21.3
decorator==4.0.11
entrypoints==0.2.2
html5lib==0.9999999
httplib2==0.10.3
httpretty==0.8.14
idna==2.5
ipykernel==4.6.1
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Markdown==2.2.0
MarkupSafe==0.23
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.13.0
oauth2==1.9.0.post1
oauthlib==2.0.2
olefile==0.44
opencv-python==3.2.0.7
pandocfilters==1.4.1
path.py==10.3.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
protobuf==3.3.0
Pygments==2.2.0
pymongo==3.4.0
python-dateutil==2.6.0
PyTumblr==0.0.6
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.17.3
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.10.0
tensorflow==0.12.0rc0
tensorflow-gpu==1.1.0
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
tweepy==3.5.0
urllib3==1.21.1
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==2.0.0
win-unicode-console==0.5
